Hack to allow propagation of Microsoft GPO entries
The certificate policy manager is configured via the Windows GPO
entries instead of our custom ADMX/ADML counterpart. We were already able
to parse these entries but we were excluding them as they didn't start
with our Ubuntu-specific prefix.

The easiest way to make this work is to add the Ubuntu prefix to the
keys that we want to use in our certificate policy manager.