Unenroll machine if cert policy is not configured
Similar to our other policy manager behaviors, if the certificate policy
is either disabled or not configured, unenroll the machine if
applicable.

To avoid running the helper Python script on every policy apply,
determine if we actually need to unenroll by checking the existence of
the Samba cache directory. Additionally, update the Python script to
remove the directory after unenrollment is successful.