Run certificate tests with clean PYTHONPATH
This handles cases where the test runner already has a PYTHONPATH set,
thus failing because of differences in the golden file output.

Ideally we would pass the PYTHONPATH via a functional option in the
manager constructor so we can maintain parallelism in tests, but I feel
this approach is a decent compromise and we don't lose much in the
process.