Add workflow to build image templates on Azure
This workflow runs on a scheduled basis and can also be triggered
manually if needed.

By default it'll figure out the currently supported Ubuntu versions and
build templates for them unless suitable templates already exist (in
other words, templates will only be built when a release transitions
from development to stable, and when no template - stable or
developmental - exists for the release at all).

This behavior can be overridden using the manual workflow inputs, which
provide a way to force refreshing the template builds, and configuring
the matrix of codenames to build.