Add scripts to provision client VM on Azure
This set of scripts is responsible for building the current working
directory of adsys as a deb package for the given codename, then
provisioning a VM from our custom VM image template, joining the
integration tests realm and installing the previously built adsys
deb(s).

The package build is done via Docker containers and tests are disabled
both to speed up the build process and to weed out false positives since
adsys package tests were not written to be run in a Docker container.

For Ubuntu releases where adsys doesn't build cleanly, such as jammy, we
provide patches to satisfy the build process.