Add barebones E2E test workflow
Add a deprovisioning script and a GitHub Actions workflow that exercises
the `provision_resources` suite which builds the ADSys DEB for all
codenames then provisions a VM on Azure, joins it to the domain,
installs ADSys, ultimately deprovisioning it.

This provides the main building block on top of which we will add
the step(s) to run the actual tests.

For now the workflow will only be triggered by hand until we decide in
what cases it's best for it to run automatically (being mindful of test
duration and Azure costs).

Fixes UDENG-1359