Rework base image check logic
Previously, the approach was to:
- always build an image if no custom image exists
- build stable image if we only have daily custom images
- never build again if a custom stable image already exists

Now, we take into consideration the build number of the Azure
Marketplace VM image, and ensure we refresh our custom images when new
base images are published. This means we have to track the build number
as part of the custom image version to ensure idempotency of the build.

Thus, the version format of our custom images becomes the following:
- X: major version, 0 for development releases, 1 for stable releases
- Y: minor version, replicates build number of the Marketplace VM, e.g.
     202310110
- Z: patch version, incremented for consecutive builds of the same minor
     version, starts at 0 (in practice this won't be encountered unless
     we force a build with the --force param)

Additionally, the behavior of the 00_check_vm_image script is to no
longer error when there is no image to be printed - this is to be able
to properly differentiate between a valid behavior and an actual error.

Now, the script always exits with 0 unless an actual error occurs. If
there is no image URN to be printed, stdout will be empty so the
consumer will know that the image for the given codename is already up
to date.