Ensure GPO URLs contain the FQDN of the controller
This is the main driver for the changes in the previous commits. The way
this behavior worked in the past is that we would use the URL returned
in the gPCFileSysPath field without any changes.

This introduced an inconsistency that went unobserved until recently.
Namely, we would get the list of GPOs using the FQDN of the domain
controller (e.g. adc.example.com), whereas the list of GPO URLs only
included the domain name (e.g. example.com). This meant that when
downloading the actual GPO data, libsmbclient would try to autodiscover
a domain controller from which to perform the download, given only the
domain name.

In some cases, especially complicated AD deployments with lots of DCs,
libsmbclient could autoresolve to an unhealthy DC (we take unhealthy to
mean any DC from which GPO files cannot be downloaded, regardless of
reason). This would fail the GPO download with a cryptic "invalid
argument" error.

Besides the chance of the above happening, autodiscovery also takes
longer as opposed to passing a valid DC FQDN to libsmbclient from the
start.

To fix this, we rewrite the GPO URL in the gPCFileSysPath field to
include the FQDN of the domain controller which essentially ensures the
DC we get the GPO list from, and the DC we download the GPO data from
are the same, minimizing the chance of mismatches like this occurring.

This has some drawbacks in the integration tests where we set up a real
SMB share and download from it, so we need to ensure the mocked server
URL is the actual SMB server.

Fixes #733 / UDENG-843