Split first run script in two parts
Not all actions in the first run script need to run on the first boot.
Some of them we can run when preparing the VM base image itself, namely:
- SSH configuration
- network configuration

The only part that needs to run on first boot is the hostname/hosts
configuration.

This will help mitigate issues with SSH not being fully initialized on
boot.