Fix race condition in client.Run
In case the executed program prints multiple lines, we could return from
the function before the scanners finished parsing the output (e.g.
running cat on a multi-line file).

This is because we were only using a mutex to synchronize access to the
combinedOutput string, and not ensuring those goroutines actually ran to
completion. We now do this with a waitgroup.