Add reboot functionality to client package
When running the tests we will need to often reboot the client VMs. Add
a helper function for this that runs the reboot command on the client,
waits for the machine to be back up again and re-establishes the SSH
connection.

Due to the reconnect, we need to specialize our struct and store more
data on it such as the configuration options and the credentials.