Add e2e tests for all policy managers
These are split in Pro and non-Pro tests, and attempt to closely mimic
actual user behavior on the system with regard to policy
application (e.g. relying on reboots for machine policy updates and
login for user policy updates).

The test assertions verify the XML policy configurations, which are also
described in the ADSys QA Plan document.

Authentication of the domain users is performed via password, we don't
encrypt this since none of the domain users have sudo rights, and the
test VMs are behind the integration tests VPN and have no public IP.

Support for some policy managers like certificates and proxy is limited
to the availability of their suggested packages, thus we can only assert
the certificates policy on mantic. Proxy is available since lunar, but
since it's soon to be EOL it's not worth to be running tests on it for a
couple of weeks.

Some helper assertions have been added on the client package for
brevity, but this has been kept to a minimum (we mostly care about
checking command output).