Add krb5 ticket detection functionality under ad
This leverages the libkrb5 API via CGo to determine the default path of
the current user's cached ticket. According to the MIT Kerberos
documentation[1] the order is:
- a prior call to krb5_cc_set_default_name()
- the KRB5CCNAME environment variable
- the default_ccache_name profile variable
- the operating system or build-time default value

The C function doesn't guarantee that the ticket is present on disk (or
that the ticket is of FILE type for that matter), so we do the
additional checks ourselves and return an error as needed.

Testing is accomplished via a C mock with a mention that integration
tests will look a bit different in the sense that we will not override
krb5_init_context for them, since that is a critical path required by
samba to work. This is not a huge loss as we do more granular tests at
the package level.

Similar to the wbclient mock, hoist the compilation logic to testutils
in order to make it accessible from the integration tests.

[1] https://web.mit.edu/kerberos/www/krb5-latest/doc/appdev/refs/api/krb5_cc_default_name.html