Get ticket path from krb5 if not present
When applying policies for the current user, if the KRB5CCNAME
environment variable is empty and the detect_cached_ticket setting is
true, try to get the default ticket path from krb5.

We don't return an error here but let the daemon handle the result, as
even if the ticket cannot be found there might still be a tracked one in
the adsys rundir.