Allow sssd backend to work without ad_domain
This restores the functionality prior to the refactor in PR #467, where
the case of having a domain section without the ad_domain setting
resorted to using the domain from the sssd.domains setting. This is
valid behavior supported and suggested[1] by sssd.

In addition to that, avoid being too lenient and still raise an error if
the domain section is empty or does not exist.

Fixes #910 / UDENG-2268

[1] https://sssd.io/docs/ad/ad-provider-manual.html#id4