Add hidden CLI command to get ticket path for user
Add a debug command under adsysctl policy responsible for returning the
default ccache path for the given user.

This command is intended to be used by the PAM module, thus we are
specializing its edge cases such as:
- configuration option not set: we do not error so PAM can proceed with
  the previous functionality (return PAM_IGNORE)
- ccache reported by libkrb5 not found on disk: we do not error since
  user can be a non-domain user

Other scenarios that can raise errors will be treated as such, and auth
will be denied.