Increase gpolist command timeout for AD tests
As the AD object is initialized multiple times in tests but only once in
production I opted for the less intuitive approach of defaulting to the
longer timeout and overriding with the production one instead of the
other way around.

This should fix the ad tests we've seen time out in armhf autopkgtests.