Customizable workflow dispatch for e2e tests
Allow the following to be overridden:
- repository (to allow testing from forks)
- branch
- codenames

Additionally, run e2e tests on pushes to main and new tags. Running on
PRs is not feasible because the jobs take around 12 minutes and we don't
want to provision Azure resources on every PR push. We leave it to the
developer to trigger ad-hoc runs once their PRs have passed the other
levels of CI and are ready for merge.