Deprovisioning can happen from two initial states
Our current model only allows for one initial state and one target
state. Some commands such as the deprovisioning one, can work
incrementally for the client/AD. For example, if we manage to provision
the client but fail to provision the AD, the script should still release
any client resources. As of now, it only works if the AD was successfully
provisioned.

There are multiple ways to handle this, but I've opted for the less
intrusive one of making states variadic arguments, where all but the
last are considered the "initial states", and the last is the state that
is being transitioned to. We still sacrifice some readability but
existing code will work, and I've made sure to document the new
intricacies of the WithStateTransition function.