Add simple retry mechanism for SSH connections
As seen in some failing CI runs, sometimes we fail to establish SSH
connection to either a client or the AD controller. I believe this is a
transient failure that could be mitigated if we retry the operation
multiple times.

This implements a simple retry mechanism that for now attempts to
connect a maximum of 10 times at an interval of 3 seconds.