Use separate variable for root SSH client
To avoid shadowing the root client and needing to make sure we have a
root connection as opposed to a user-specific one, use a separate
variable to make the distinction.

This makes the scope of the access easier to predict and understand.

By doing this we can isolate log collection to standalone defer calls.