Add option to allow multiple coverage formats
This will allow us to call the `TrackTestCoverage` function multiple
times per test, optionally specifying a coverage format that will be
used to give preferrential treatment to some coverage reports.

The main purpose of the functionality is to differentiate between
coverage formats like Go and XML which aren't meant to be used together.