Use nolint instead of nosec for gosec
The nosec directive is broken as of gosec 2.19 -- we can work around
this by using the nolint directive from golangci-lint. The main drawback
is that this disables the entire linter, not just one rule, but I still
kept the rule name for reference.

After this is merged we can bump golangci-lint to 1.56.2.