Fix argument for az sig image-version create
It appears there's been a regression in version 2.59.0 of the Azure CLI,
where az sig image-version create fails when being passed the
--managed-image flag, see the related failing CI run,[1] and the reported
issue on the azure-cli project.[2]

Because of this we are not able to create new versions for any of our VM
templates.

To fix this we can apply the workaround from the linked issue and pass
the VM ID inside the `--virtual-machine` argument.

[1] https://github.com/ubuntu/adsys/actions/runs/8677801641/job/23793927248
[2] https://github.com/Azure/azure-cli/issues/28700