Disable systemd-networkd-wait-online on Noble
We already did this for Jammy and it appears that Noble exhibits the
same issue which causes provisioning to take 2 minutes longer than usual
(due to the timing out).

What's surprising to me is that we didn't need to do the same for Lunar
or Mantic and I'm afraid I don't have a logical explanation for this,
but hey... it works!