Remove XML declaration from glib schemas
This is a workaround so we don't fail[1] when parsing schemas with invalid
XML declarations, such as org.gnome.Extensions.gschema.xml.[2]
Interestingly this doesn't seem to bother glib-compile-schemas but fails
with most XML parsers (including Go's parser).

While I've already fixed this upstream,[3] it'll take a while until the
updated version feeds into Ubuntu so we need a workaround within adsys
too, otherwise we are blocked from refreshing the ADMX/ADML files.

For simplicity I've opted to remove the XML declaration line if it
occurs, this way we don't concern ourselves checking for specific syntax
errors.

[1] https://github.com/ubuntu/adsys/actions/runs/8809678726/job/24180895123
[2] https://gitlab.gnome.org/GNOME/gnome-shell/-/blob/41b94ccf11a66c5b315f834061cdec82a622f4ec/subprojects/extensions-app/data/org.gnome.Extensions.gschema.xml
[3] https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3281