Adjust bubbletea tests after dependency update
As part of the latest bubbletea release, there was an optimization in
batches with a single command[1]. This ended up breaking our tests
because we inferred the batchMsgType from a Batch call with a single
message, which after the optimization is no longer a tea.BatchMsg, but a
plain struct.

To fix this we could either pass another Cmd to the tea.Batch call, or
simply do the reflection inline since we know our candidates should be
of type tea.BatchMsg. I've opted for the latter which seems cleaner.

[1] https://github.com/charmbracelet/bubbletea/pull/875/commits/2d82d8ed94b3d8386738a7b8e97e90e52199548b