Add TiCS cron job workflow
Because the TiCS suite takes quite a long time to install and run, our
best bet is to run it on a nightly basis as a scheduled job. I've
selected 4 AM UTC as the start time, this should be a quiet as no one is
supposed to be working at this hour on any of our team's timezones (I
hope), so no risk of changes coming through while the workflow is
running.

To set up TiCS itself, we follow TIOBE's guidelines, namely:
- place the coverage in the expected location
- install TiCS
- compile all of our binaries
- run TiCS
- archive and upload logs as artifacts