Change coverage publish to nightly prerelease
In short, this saves us from an additional test suite run in our
forthcoming TiCS scheduled workflow.

Update the coverage report to be uploaded as a pre-release artifact on
every 'main' commit -- the consumer can then simply perform a request
like the following to download the file:

https://github.com/ubuntu/adsys/releases/download/nightly/Cobertura.xml

A caveat of this approach is that the tag has to exist, and the ref the
tag is pointing to will not change. Unfortunately there's no one size
fits all solution for this that:
- handles creating the tag
- keeps the ref up to date
- replaces the coverage asset with the latest

I think these are reasonable drawbacks since tag creation is a one-time
action, and we don't (yet) need the tag to point to the latest ref.