Add script to destroy old Azure image versions
While checking on our Azure costs I noticed that Gallery costs were
slightly increasing over time. This happens because we are charged per
VM image version and up until now we had no policy to clean up unused
versions.

To mitigate this I've devised a script intended to be run just after the
image finishes building (as part of the weekly image build workflow).

The script destroys the oldest N-2 image versions of the specified
codename by default, with the number of versions to keep being
configurable via CLI argument in case we ever need to run it by hand.

A future commit will integrate this step as part of the scheduled
workflow.