Skip tests on Oracular until images are available
Currently e2e tests are failing on Oracular because there's no Docker
image for it yet so we cannot build the adsys deb. On top of this, we
don't have an Azure VM template for it either.

We should revisit this in a month or two as the new version stabilizes.