Reverse case logic in adsysd main
This is hard to explain, but... it's a trick to be able to exercise the
main function from within tests without the logic defaulting to starting
the client and ultimately failing because there's no socket to connect
to. Because `os.Args[0]` is occupied by the Go test runner we can't alter
it, so defaulting to starting the daemon is the next best thing.