Remove broken prerm purge stanza
It was pointed to us that dpkg NEVER calls a prerm script with an
argument of 'purge'; 'purge' is only ever given as an argument to a
postrm script,[1] so this bit of code never actually worked.

Interestingly hundreds of packages in the wild fall in this pitfall[2],
so we are not the only victims here.

I'm choosing to remove the stanza altogether since there's no (simple)
way for us to execute `adsysctl` in a purge context, and I don't
consider having this working to be a huge win for the project.

Because this functionality was never documented properly in the
changelog (the full changelog entry is: "Add adsysctl policy purge
command to purge applied policies"), no actual public-facing promises
were made, so I don't see the point in us documenting the removal of
this hiccup either.

[1] https://wiki.debian.org/MaintainerScripts
[2] https://github.com/search?q=purge+path%3Adebian%2F+path%3A**%2F*prerm&type=Code