Archive build.log as part of the TICS logs
This is where the logs pertaining to the adsys build are saved. If
something happens to fail during project build (in this case because
libpam0g-dev was not installed) nothing visible is logged in the other
logfiles so be sure to also archive this one to save us some headaches.