Add e2e scripts to collect PAM module coverage
Add two new scripts, one that recompiles the PAM module to add coverage
support, and one that collects the coverage at the end of the suite,
before deprovisioning.

The coverage is collected in Cobertura format and downloaded to the
local runner, which we can then upload as a artifact to be used with
other workflows such as the TICS check.