Update e2e tests workflow to collect PAM coverage
For simplicity we only do these additional steps on a single runner,
namely the Noble one. It's not worth it attempting to pool the coverage
from all runners since it is most likely identical.

This adds the burden of maintenance to manually update the codename
whenever Noble is no longer desirable, but it'll likely be years before
that is required.

The collected coverage is uploaded as an artifact on the workflow, with
a deterministic name, so it can be easily downloaded and reused in
different workflows.