Filter out releases with no VM images
Rather than have the e2e tests failing constantly until an image becomes
available, we can be smarter about it and dynamically exclude the
codename until the Azure CLI reports an existing image.

Of course, this doesn't directly map to the existence of a customized
image built by our "E2E - Build image templates" workflow -- but it
gives a clue as to when it becomes possible to build such an image.