Disable unattended-upgrades in e2e templates
While we love keeping our system up to date, in the e2e test scenarios
it's more of a hindrance because it could run at the same time as we are
running 'apt install', which would fail our run.

This is one of the most common causes of flakiness in our e2e suite.