Update E2E matrix script to store Ubuntu versions
Now that the Ubuntu version is a hard requirement for the 00_check_image
script, we need access to it in our GitHub Actions workflow.

I've updated the logic to manipulate the output of `distro-info
--fullname` which looks like the following:

Ubuntu 20.04 LTS "Focal Fossa"
Ubuntu 22.04 LTS "Jammy Jellyfish"
Ubuntu 23.10 "Mantic Minotaur"
Ubuntu 24.04 LTS "Noble Numbat"
Ubuntu 24.10 "Oracular Oriole"

It's not particularly suitable for parsing, but it's the only way we can
manipulate distro-info into giving us both the codename and version on
the same line.

We save the version in the job output, and re-use it when executing the
Go script.