Take 4 characters off the unique e2e hostname
When creating AD users for testing we use the codename to which we
append a suffix like '-usr' or '-adm'. Due to oracular being quite a
long name added up to the 8 character unique hex, AD fails to create the
user since it ends up being above 20 characters long.

    New-ADUser : The name provided is not a properly formed account name

There's many ways to fix this, but I'm doing it at a template
level (meaning a refresh will be required), essentially halving the
unique part from 8 characters to 4 characters. We don't provision lots
of resources so I'm not expecting clashes due to the unique part being
shorter.