Missing requirement.txt on RtD
build_requirements.py expects requirements.txt to be at the certain
location. Since the documentation is in a subdirectory, it is not where
build_requirements.py would like it to be (missing ./docs)
We build the path of requirements.txt dynamically so it is always in the
same directory than the python script.