Standardize runner across CI
ubuntu-latest is still at 22.04 on github (see
https://github.com/actions/runner-images). Since we rely on polkit for
some policies and there was a big set of changes on the package between
22.04 and 24.04, we need to bump the runners to match the docker
container that we use for the integration tests. To do that, it's better
to use repo variables to control the runner so that we can switch back
to ubuntu-latest once it moves on to 24.04.