Fix username parsing in PAM's slash_to_at_username
This function is supposed to transform a domain.com\user username to the
@ format (i.e. user@domain.com). However, it was leaving the parsed
username with bad characters in the end (maybe an off behavior from
strncpy or something wrong in our string manipulation). This takes a
more "direct" approach by selecting the precise substring that we want
to append to the parsed name.

Co-authored-by: Didier Roche-Tolomelli <didrocks@ubuntu.com>