Bump Go to 1.23
We now rely on Go >= 1.22.7, but Jammy and Noble versions (for the
golang-1.22-go package) only go up to 1.22.2, which means we would be
unable to build for those distributions. Since those versions now also
have golang-1.23-go available, let's use that package instead and also
bump the version here to ensure we have CI coverage.