Fix distro patches for e2e tests
After adsys#1175, we bumped the minimum Go version to 1.23. On a
toolchain level, this should not make such a difference since Go is able
to update itself as needed according to what's specified in go.mod.
However, the PR also introduced changes to d/control and those broke the
patches that we need to apply in order to run the e2e tests.