Update d/rules to use the local toolchain
If the toolchain version is higher than the current Go version, Go will
try to download the higher version during the build time. This is an
issue on Launchpad builds, since the builders there do not have access
to internet. To avoid that, let's make sure that d/rules overrides this
behavior and forces Go to use the version available in the archive.