Update flow for parsing certificate templates
The previous implementation tried to autodiscover the CEPCES server to
which it should connect to and fetch the templates for a specific CA.
However, it was setup in a way which would always override what was
configured in the cepces config file.

This was problematic as it was always assuming that the CEPCES server
was the same as the CA source, which isn't always the case.

These changes now allow us to override this behavior through the
cepces.conf file or keep it as is, provided that the configuration file
is the default one.

This also improves the output and error handling of the process, as it
didn't return any kind of message when failing to parse the templates
properly.