Update .golangci configuration with latest standards
We were using some deprecated linters and configuration settings, so
it's about time that we updated them to the current format. Also, adsys
was still using separate config files for local development and CI while
all of our other projects unified them into a single config, so let's
fix this as well.