Change argument from io.Reader to os.File
In order to be able to resize the buffer later, we need to have control
over the file cursor otherwise it would move forward and we would not be
able to retry reading the large entry.

Sadly, this implies on removing the fuzz test as we can't use os.File as
argument.