Allow searching GPO list by userPrincipalName
Users normally login using their user principal name and also usually
the sAMAccountName is the same user removing the @ and the domain.
Currently the GPO search is done by removing the @ and the rest from the
login name and searching that value as sAMAccountName.
This fails for users where their sAMAccountName is not a suffix of their
userPrincipalName. This patch adds a search using the userPrincipalName
in case the search for sAMAccountName fails to support users with that
configuration.

Fixes #1273

Signed-off-by: Jorge Merlino <jorge.merlino@canonical.com>